Fixed many, MANY, boss bugs
- Updated script to properly use enemy refactor code
- Fixed AI being slow and weighty
- Reduced player iframes by half
- Adjusted all enemy ai sizes (should help them pathfind)
- Boss animator now properly tracks can attack states (previously allowed for faster than intended attack cooldowns)
- Refactored boss damage code in general to use OverlapCircleAll instead of toggling a collider

+++ GMTK-2025/Assets/Scripts/Utilities/Saving.cs
@@ -20,7 +20,7 @@ public class Saving : MonoBehaviour
         public float maxSpeed;
         public float health;
         public float maxHealth;
-        public int invincibilityFrames;
+        public float invincibilityFrames;
         public float experience;
         public float castStrength;
         public float castSpeed;
